requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
google-generativeai
langchainhub
langchain-google-genai
langchain-cli
numexpr
groq
langchain-groq
beautifulsoup4